backend-python/requirements.txt
fastapi
uvicorn
beanie
motor
python-dotenv
pydantic
python-multipart
